Updating clarity of Document.sol by YC
Variables:
Public Variables: We have ensured that title, url, dochash, signature, signer, owningContract, and timestamp are public, which automatically generates getter functions for easy access and improved transparency.

Constructor:
Initialization: The constructor is meticulously designed for initializing the contract. It accepts parameters like _owningContract, _signer, _signature, _dochash, _url, and _title, setting the foundation of the contract with robust error handling.

Functions:
addDocument: This function is integral for adding new documents. It is carefully crafted to handle inputs like signer, title, url, docHash, and signature. To enhance user experience and debugging, we've introduced a revert statement with a detailed error message in case of duplicate URL entries.
getDocuments: An external function that returns an array of MembershipManagement.Doc[]. This feature is pivotal in fetching all document information efficiently and accurately.
Overall Enhancements and Recommendations:
Error Handling in Constructor: We have implemented the use of revert along with customized error messages in the Document contract's constructor. This approach significantly increases code readability and simplifies debugging.
Event Logging: In the DocumentsHolder contract, we are considering the addition of events to log the addition of documents, thereby increasing the traceability and transparency of operations.

Comments and Annotations: To further improve code readability and maintenance, I have added comprehensive comments and annotations throughout both contracts.

Access Control: Ensured appropriate access controls were in place for all public and external functions, aligning with best practices in smart contract development.